# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Librarian API",
    "version": "0.1.0",
    "description": "Local-first API for EPUB ingestion, local embedding maintenance, vector search, and library inspection."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Default local development server"
    }
  ],
...
  "paths": {
...
    "/ingestion/run": {
      "post": {
        "tags": ["ingestion"],
        "operationId": "runIngestion",
        "summary": "Run EPUB ingestion",
        "description": "Scans EPUB files, parses text, chunks books, stores book and chunk rows, optionally generates embeddings, and can enqueue asynchronous summary jobs for newly ingested books.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionRunRequest"
              },
              "examples": {
                "ingestOnly": {
                  "summary": "Ingest EPUBs without embeddings",
                  "value": {
                    "books_dir": "/books",
                    "database_url": "sqlite:///data/librarian.db",
                    "list_epubs": true
                  }
                },
                "ingestAndEmbed": {
                  "summary": "Ingest and embed with Ollama",
                  "value": {
                    "books_dir": "/books",
                    "database_url": "sqlite:///data/librarian.db",
                    "force": false,
                    "embed_chunks": true,
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm",
                    "ollama_base_url": "http://host.docker.internal:11434",
                    "embedding_batch_size": 16
                  }
                },
                "ingestAndQueueSummaries": {
                  "summary": "Ingest and queue summary jobs",
                  "value": {
                    "books_dir": "/books",
                    "database_url": "sqlite:///data/librarian.db",
                    "enqueue_summaries": true,
                    "summary_generation_provider": "codex",
                    "summary_generation_model": "codex",
                    "summary_detail": "medium"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingestion run summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionRunResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/ingestion/summary": {
      "get": {
        "tags": ["ingestion"],
        "operationId": "getIngestionSummary",
        "summary": "Get ingestion database summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/DatabaseUrl"
          }
        ],
        "responses": {
          "200": {
            "description": "Counts and status totals for the ingestion database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/ingestion/status": {
      "get": {
        "tags": ["ingestion"],
        "operationId": "getIngestionStatus",
        "summary": "Get ingestion pipeline status",
        "description": "Returns UI-oriented progress for chunking, summarizing, and tagging so clients can poll while ingestion-related background work runs.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DatabaseUrl"
          }
        ],
        "responses": {
          "200": {
            "description": "Current ingestion pipeline progress.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionPipelineStatus"
                },
                "examples": {
                  "partialProgress": {
                    "value": {
                      "database_url": "sqlite:///data/librarian.db",
                      "total_books": 2,
                      "chunking": {
                        "status": "complete",
                        "total_books": 2,
                        "completed_books": 2,
                        "pending_books": 0,
                        "running_books": 0,
                        "failed_books": 0,
                        "percent_complete": 100.0,
                        "details": {
                          "ingested_books": 2,
                          "total_chunks": 42,
                          "total_chunk_duration_seconds": 3.42,
                          "avg_chunk_duration_seconds": 1.71
                        },
                        "active_jobs": []
                      },
                      "summarizing": {
                        "status": "in_progress",
                        "total_books": 2,
                        "completed_books": 1,
                        "pending_books": 0,
                        "running_books": 1,
                        "failed_books": 0,
                        "percent_complete": 50.0,
                        "details": {
                          "book_summaries": 1,
                          "chapter_summaries": 12,
                          "summary_jobs_pending": 0,
                          "summary_jobs_running": 1,
                          "summary_jobs_completed": 1,
                          "summary_jobs_failed": 0,
                          "total_summary_duration_seconds": 420.5,
                          "avg_summary_duration_seconds": 420.5,
                          "max_summary_duration_seconds": 420.5,
                          "unqueued_books": 0
                        },
                        "active_jobs": [
                          {
                            "job_id": "summary-job-2",
                            "book_id": "book-2",
                            "relative_path": "Forward the Foundation.epub",
                            "title": "Forward the Foundation",
                            "authors": ["Isaac Asimov"],
                            "provider": "ollama",
                            "model": "llama3.2:3b",
                            "detail": "medium",
                            "attempts": 1,
                            "stage": "chapter",
                            "current": 4,
                            "total": 18,
                            "message": "Generating summary for Chunks 24-31.",
                            "started_at": "2026-06-23T11:58:00+00:00",
                            "duration_seconds": 120.0,
                            "updated_at": "2026-06-23T12:00:00+00:00"
                          }
                        ]
                      },
                      "tagging": {
                        "status": "in_progress",
                        "total_books": 2,
                        "completed_books": 1,
                        "pending_books": 2,
                        "running_books": 1,
                        "failed_books": 0,
                        "percent_complete": 50.0,
                        "details": {
                          "books_with_tags": 1,
                          "books_with_genres": 1,
                          "total_tags": 8,
                          "total_genres": 2,
                          "metadata_jobs_pending": 1,
                          "metadata_jobs_running": 1,
                          "metadata_jobs_completed": 1,
                          "metadata_jobs_failed": 0,
                          "total_metadata_duration_seconds": 38.4,
                          "avg_metadata_duration_seconds": 38.4,
                          "max_metadata_duration_seconds": 38.4
                        },
                        "active_jobs": [
                          {
                            "job_id": "metadata-job-1",
                            "book_id": "book-2",
                            "relative_path": "Forward the Foundation.epub",
                            "title": "Forward the Foundation",
                            "authors": ["Isaac Asimov"],
                            "job_type": "genres",
                            "source_summary_provider": "ollama",
                            "source_summary_model": "llama3.2:3b",
                            "source_summary_detail": "medium",
                            "provider": "ollama",
                            "model": "llama3.2:3b",
                            "attempts": 1,
                            "stage": "metadata",
                            "current": 0,
                            "total": 1,
                            "message": "Generating genres metadata.",
                            "updated_at": "2026-06-23T12:01:00+00:00"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/embeddings/rebuild": {
      "post": {
        "tags": ["embeddings"],
        "operationId": "rebuildEmbeddings",
        "summary": "Rebuild stored chunk embeddings",
        "description": "Regenerates embeddings from existing chunk rows without deleting books, chunks, or raw text.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebuildEmbeddingsRequest"
              },
              "examples": {
                "currentModel": {
                  "summary": "Rebuild selected model",
                  "value": {
                    "database_url": "sqlite:///data/librarian.db",
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm",
                    "reset": true
                  }
                },
                "benchmarkModel": {
                  "summary": "Clear all embeddings before benchmarking a model",
                  "value": {
                    "database_url": "sqlite:///data/librarian.db",
                    "embedding_provider": "ollama",
                    "embedding_model": "nomic-embed-text",
                    "ollama_base_url": "http://host.docker.internal:11434",
                    "batch_size": 8,
                    "chunk_page_size": 250,
                    "reset_all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Embedding rebuild summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RebuildEmbeddingsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/embeddings/query": {
      "post": {
        "tags": ["embeddings"],
        "operationId": "embedQuery",
        "summary": "Embed a user query",
        "description": "Creates an embedding vector for direct inspection or for clients that need query vectors separately from search.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedQueryRequest"
              },
              "examples": {
                "ollama": {
                  "summary": "Embed with Ollama",
                  "value": {
                    "query": "books about memory and identity",
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm"
                  }
                },
                "noop": {
                  "summary": "Shape test with no-op provider",
                  "value": {
                    "query": "clockwork gardens",
                    "embedding_provider": "noop"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query embedding vector and metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedQueryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/search": {
      "post": {
        "tags": ["retrieval"],
        "operationId": "searchChunks",
        "summary": "Search stored chunk embeddings",
        "description": "Embeds a user query, loads stored chunk embeddings for the same provider/model, scores them with cosine similarity, and returns ranked chunks with book metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              },
              "examples": {
                "default": {
                  "summary": "Search with the default local model",
                  "value": {
                    "query": "what does the author say about memory?",
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm",
                    "limit": 5
                  }
                },
                "authorFiltered": {
                  "summary": "Search within books by one author",
                  "value": {
                    "query": "How brutal and terrible is war?",
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm",
                    "author": "Erich Maria Remarque",
                    "limit": 10
                  }
                },
                "explicitDatabase": {
                  "summary": "Search an explicit local database",
                  "value": {
                    "query": "fantasy books with political intrigue",
                    "database_url": "sqlite:///data/librarian.db",
                    "embedding_provider": "ollama",
                    "embedding_model": "all-minilm",
                    "ollama_base_url": "http://host.docker.internal:11434",
                    "limit": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked chunk matches.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "IngestionRunRequest": {
        "type": "object",
        "properties": {
          "books_dir": {
            "type": ["string", "null"],
            "description": "EPUB source directory. Defaults to LIBRARIAN_BOOKS_DIR."
          },
          "database_url": {
            "type": ["string", "null"],
            "description": "Storage URL. Defaults to LIBRARIAN_DATABASE_URL.",
            "example": "sqlite:///data/librarian.db"
          },
          "force": {
            "type": "boolean",
            "default": false,
            "description": "Re-parse unchanged EPUB files."
          },
          "list_epubs": {
            "type": "boolean",
            "default": false,
            "description": "Include discovered EPUB metadata in the response."
          },
          "embed_chunks": {
            "type": "boolean",
            "default": false,
            "description": "Generate embeddings for chunks created during this ingestion run."
          },
          "embedding_provider": {
            "type": ["string", "null"],
            "enum": ["noop", "ollama", null]
          },
          "embedding_model": {
            "type": ["string", "null"],
            "example": "all-minilm"
          },
          "ollama_base_url": {
            "type": ["string", "null"],
            "example": "http://host.docker.internal:11434"
          },
          "embedding_batch_size": {
            "type": "integer",
            "default": 16,
            "minimum": 1
          },
          "enqueue_summaries": {
            "type": "boolean",
            "default": false,
            "description": "Queue asynchronous chapter/book summary jobs for newly ingested books."
          },
          "summary_generation_provider": {
            "type": ["string", "null"],
            "enum": ["noop", "ollama", "codex", null]
          },
          "summary_generation_model": {
            "type": ["string", "null"],
            "example": "codex"
          },
          "summary_detail": {
            "type": "string",
            "enum": ["short", "medium", "detailed"],
            "default": "medium"
          }
        }
      },
...
    }
  }
}